updates routing structiure
